refactor: moved types into index
